fix(data): bind embedding BLOB via ByteArray; FloatArray param never worked
Discovered while writing HybridSearchEngineReadPathTest for F-G
read-side: dao.updateEmbedding(id, embedding: FloatArray) generated
malformed SQL on the unit test DB:

  UPDATE notifications SET embedding = ?,?,?,?,?,... WHERE id = ?

Room expands FloatArray query parameters as if they were
Iterable<Float>, generating one ? per element. The TypeConverter in
Converters.kt (FloatArray<->ByteArray) is only used by Room when
serializing entity FIELDS, not raw @Query parameters.

That means BOTH production call sites of this DAO method have been
silently failing since the embedding column was added:
  - DatabaseMigrator.embedAll (v18 one-shot backfill at upgrade time)
  - NotificationLoggerService.indexEmbeddingForSingle (commit f279324
    capture-side wiring)

Result: every notification has had embedding=NULL in practice. The
HybridSearchEngine semantic leg had nothing to score. FTS4 leg worked
but semantic search was dead.

Fix: change DAO method signature to ByteArray (bindable), update both
call sites to convert via Converters.fromFloatArray at the boundary so
the byte layout matches what the entity TypeConverter reads back via
Converters.toFloatArray. Same conversion shape as the entity column
uses, so writes and reads round-trip identically.

Verified:
  - ./gradlew :app:compileDebugKotlin -> BUILD SUCCESSFUL
  - ./gradlew :app:testDebugUnitTest -> 145 -> 149 tests, 0 failures,
    1 skipped (the pre-existing @Ignore on MigrationTest).

Note on test architecture: this fix also unblocks the F-G read-side
wire-up (HybridSearchEngine integration into the search box). The next
commit wires HybridSearchEngine into LogHistoryScreen and
ActiveNotificationsScreen; the one after that updates the audit doc
to reflect that F-G is now fully resolved.

+++ app/src/main/java/com/jeffers/notimindlite/data/local/NotificationDao.kt
@@ -115,7 +115,7 @@ abstract class NotificationDao {
     abstract suspend fun getNotificationsNeedingVectorization(): List<NotificationEntity>
 
     @Query("UPDATE notifications SET embedding = :embedding WHERE id = :id")
-    abstract suspend fun updateEmbedding(id: Long, embedding: FloatArray)
+    abstract suspend fun updateEmbedding(id: Long, embedding: ByteArray)
 
     @Query("SELECT * FROM notifications WHERE isPinned = 1 ORDER BY postTime DESC")
     abstract fun getPinnedNotificationsFlow(): Flow<List<NotificationEntity>>
@@ -231,7 +231,7 @@ abstract class NotificationDao {
     abstract suspend fun updateSyncStatusBatch(keys: List<String>, status: SyncStatus, syncedAt: Long = System.currentTimeMillis())
 
     @Transaction
-    open suspend fun updateEmbeddingsBatch(items: List<Pair<Long, FloatArray>>) {
+    open suspend fun updateEmbeddingsBatch(items: List<Pair<Long, ByteArray>>) {
         for ((id, embedding) in items) {
             updateEmbedding(id, embedding)
         }
